requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-ninja==1.4.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
kombu==5.5.3
MarkupSafe==3.0.2
phonenumbers==9.0.5
pillow==11.2.1
prompt_toolkit==3.0.51
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
gunicorn
psycopg2-binary
redis==6.1.0
sendgrid
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug
whitenoise==6.9.0
dj-database-url
django-redis
